+++ doorstop/__init__.py
@@ -1,7 +1,7 @@
 """Package for doorstop."""
 
 __project__ = 'Doorstop'
-__version__ = '0.5-dev'
+__version__ = '0.6-dev'
 
 CLI = 'doorstop'
 GUI = 'doorstop-gui'
@@ -10,6 +10,6 @@ VERSION = __project__ + '-' + __version__
 try:
     from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
     from doorstop.core import Item, Document, Tree
-    from doorstop.core import build, report, find_document, find_item
+    from doorstop.core import importer, build, report, find_document, find_item
 except ImportError:  # pragma: no cover, manual test
     pass